pyproject.toml
[project]
name = "gridiron-ngs"
version = "0.1.0"
description = "NFL NGS Data Processing Pipeline"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "polars>=0.20.0",
    "pyarrow>=14.0.0",
    "numpy>=1.26.0",
    "pyyaml>=6.0",
    "streamlit>=1.50.0",
    "plotly>=6.5.0",
    "tqdm>=4.67.1",
]

